+++ tests/testthat/test-checks_df.R
@@ -1,3 +1,12 @@
+examples_path <- system.file("example_packages", package = "checked")
+path <- c(
+  test_path("testing_pkgs", "DALEXtra"),
+  test_path("testing_pkgs", "rd2markdown"),
+  file.path(examples_path, "exampleGood"),
+  file.path(examples_path, "exampleOkay"),
+  file.path(examples_path, "exampleBad")
+)
+
 test_that("rev_dep_check_tasks_df works with deafult params", {
   expect_silent(
     df <- rev_dep_check_tasks_df(
@@ -26,6 +35,80 @@ test_that("rev_dep_check_tasks_df works with deafult params", {
   expect_no_error(print(df$custom))
 })
 
+test_that("task_df functions can specify subprocesses configuration", {
+  expect_silent(
+    df <- rev_dep_check_tasks_df(
+      test_path("testing_pkgs", "DALEXtra"),
+      repos = "https://cran.r-project.org/",
+      env = c("NOT_CRAN" = "false", options::opt("check_envvars")),
+      args = c("--some-option", "--other-option", options::opt("check_args")),
+      build_args = c("--yet-another-option", options::opt("check_build_args"))
+    )
+  )
+  
+  expect_identical(
+    df$package[[1]]$env, 
+    c("NOT_CRAN" = "false", options::opt("check_envvars"))
+  )
+  expect_identical(
+    df$package[[1]]$args, 
+    c("--some-option", "--other-option", options::opt("check_args"))
+  )
+  expect_identical(
+    df$package[[1]]$build_args, 
+    c("--yet-another-option", options::opt("check_build_args"))
+  )
+
+  expect_silent(
+    df <- rev_dep_check_tasks_df(
+      test_path("testing_pkgs", "DALEXtra"),
+      repos = "https://cran.r-project.org/",
+      env = c("NOT_CRAN" = "false"),
+      args = c("--some-option", "--other-option"),
+      build_args = c("--yet-another-option")
+    )
+  )
+  
+  expect_identical(
+    df$package[[1]]$env, 
+    c("NOT_CRAN" = "false")
+  )
+  expect_identical(
+    df$package[[1]]$args, 
+    c("--some-option", "--other-option")
+  )
+  expect_identical(
+    df$package[[1]]$build_args, 
+    "--yet-another-option"
+  )
+  
+  withr::with_envvar(
+    c(R_CHECKED_CHECK_ARGS = "--some-option --other-option --another-option",
+      R_CHECKED_CHECK_BUILD_ARGS = "--yet-another-option"), {
+        expect_silent(
+          df <- source_check_tasks_df(
+            path,
+            build_args = c("--another-option", options::opt("check_build_args"))
+          )
+        )
+      }
+  )
+  
+  expect_identical(
+    df$package[[1]]$env, 
+    options::opt("check_envvars")
+  )
+  expect_identical(
+    df$package[[1]]$args, 
+    c("--some-option", "--other-option", "--another-option")
+  )
+  expect_identical(
+    df$package[[1]]$build_args, 
+    c("--another-option", "--yet-another-option")
+  )
+  
+})
+
 test_that("rev_dep_check_tasks_df development_only = TRUE", {
   expect_silent(
     df <- rev_dep_check_tasks_df(
@@ -51,17 +134,8 @@ test_that("rev_dep_check_tasks_df development_only = TRUE", {
 })
 
 test_that("source_check_tasks_df works as expected", {
-  examples_path <- system.file("example_packages", package = "checked")
   expect_silent(
-    df <- source_check_tasks_df(
-      c(
-        test_path("testing_pkgs", "DALEXtra"),
-        test_path("testing_pkgs", "rd2markdown"),
-        file.path(examples_path, "exampleGood"),
-        file.path(examples_path, "exampleOkay"),
-        file.path(examples_path, "exampleBad")
-      )
-    )
+    df <- source_check_tasks_df(path)
   )
   expect_s3_class(df, "data.frame")
   expect_equal(NROW(df), 5)
@@ -79,23 +153,16 @@ test_that("source_check_tasks_df works as expected", {
 })
 
 test_that("source_check_tasks_df aliases are properly handled", {
-  examples_path <- system.file("example_packages", package = "checked")
-  names <- c("DALEXtra_new", "rd2markdown_new", "exampleGood_new", "exampleOkay_new", "exampleBad_new")
-  path <- c(
-    test_path("testing_pkgs", "DALEXtra"),
-    test_path("testing_pkgs", "rd2markdown"),
-    file.path(examples_path, "exampleGood"),
-    file.path(examples_path, "exampleOkay"),
-    file.path(examples_path, "exampleBad")
-  )
-  names(path) <- names
-
+  broken_names <- c("DALEXtra_new", "rd2markdown_new", "exampleGood_new", "exampleOkay_new", "exampleBad_new")
+  path_broken <- path
+  names(path_broken) <- broken_names
+  
   expect_silent(
-    df <- source_check_tasks_df(path)
+    df <- source_check_tasks_df(path_broken)
   )
 
   expect_true(all(endsWith(df$alias, "_new")))
-  expect_equal(df$alias, names)
+  expect_equal(df$alias, broken_names)
 
   expect_silent(
     df <- source_check_tasks_df(c(